pyproject.toml
[project]
name = "travel-assistant-langgraph"
version = "1.0.0"
description = "Enterprise-grade Travel Assistant using LangGraph, Gemini API, and FastAPI with comprehensive logging and production features"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Chitti Vijay", email = "chittivijay2003@example.com"}
]
keywords = ["ai", "travel", "assistant", "langgraph", "gemini", "fastapi", "enterprise"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "google-generativeai>=0.3.2",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langchain-core>=0.1.10",
    "langgraph>=0.0.20",
    "python-dotenv>=1.0.0",
    "nest-asyncio>=1.6.0",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
    "slowapi>=0.1.9",
    "cachetools>=5.3.2",
    "tiktoken>=0.5.2",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=html --cov-report=term-missing"
